feat: strategy-defined ledger features (pathTrie k-grams, hcb edge-buckets)
A coverage strategy can now publish the vocabulary the mutation pool
accounts feature ownership in, instead of the ledger always falling back
to bare edge indices:

- CoverageEngine gains an optional features closure, collected inside
  the same gated window as an accepting decide.
- .pathTrie publishes sliding k-grams (default k=2, configurable via
  .pathTrie(gramLength:)) of the ordered first-hit path — PathTrie now
  records the path and judges-and-collects in one critical section.
  Gram hashes are deterministic position-dependent FNV-1a (PathGrams).
- .hitCountBuckets publishes (edge << 8 | bucketBit) pairs.
- Features widen to UInt64 end-to-end (ledger, entropic policy + math);
  PoolIterationOutcome.resolvedFeatures is the single vocabulary every
  pool component reads, falling back to widened edge indices.

Rationale: pathTrie accepts on path novelty but the pool culled on edge
sets, capping retention at one entry per edge (142 on fsub). K-grams
match the retained diversity to the acceptance criterion (measured: 1.1k
features at k=2, decaying accept rate vs pathTrie's flat 37%).

+++ Sources/PropertyTestingKit/Fuzzing/Plugins/FuzzPluginHandler.swift
@@ -614,8 +614,8 @@ struct EntropicRarityTerms {
 /// keep eliciting rare features hold energy, while the abundance term decays
 /// seeds that execute without yielding.
 func entropicYieldRarityTerms(
-    yield: [UInt32: Int],
-    globalFreqs: [UInt32: Int],
+    yield: [UInt64: Int],
+    globalFreqs: [UInt64: Int],
     rareFeatureThreshold: Int
 ) -> EntropicRarityTerms {
     var energy = 0.0
@@ -635,8 +635,8 @@ func entropicYieldRarityTerms(
 /// Compute an entry's rarity terms from the current global frequencies.
 /// Called at acceptance time (frequencies only change then), not per drain.
 func entropicRarityTerms(
-    features: [UInt32],
-    globalFreqs: [UInt32: Int],
+    features: [UInt64],
+    globalFreqs: [UInt64: Int],
     rareFeatureThreshold: Int
 ) -> EntropicRarityTerms {
     var energy = 0.0
@@ -694,9 +694,9 @@ func entropicWeightCombining(
 /// the plugin's hot path uses the split form, and the equivalence test holds
 /// the two together.
 func entropicWeight(
-    features: [UInt32],
+    features: [UInt64],
     mutations: Int,
-    globalFreqs: [UInt32: Int],
+    globalFreqs: [UInt64: Int],
     totalRareFeatures: Int,
     totalMutations: Int,
     corpusSize: Int,
